Remove FUSE landcover commands and clean up rootFrac
Previously, the code included landcover commands as an experiment related to the FUSE project that have been removed.

The alternative option for setting root fractions has been cleaned up. If rootFrac in [OPTIONS] is set to False, then the root distribution in a soil layer is equal to its relative contribution to the soil column. For example, if soil layer 1 is 10% of the soil column, then 10% of the roots are in soil layer 1. In the settings file, rootFrac has been moved into the [OPTIONS] section.