Change bilinear Interpolation method to work at boundaries of meteo input map
the bilinear interpolation method always needs to set a buffer around the maskmap to get consistent interpolation results. This works fine as long as the MaskMap (the modelled extent) is smaller than the meteo input. If the MaskMap includes the boundary of the meteo input, no data for the buffer exists. In this case, an artifical buffer needs to be constructed by repeating the same row/column. This is a rare occasion if you use global input data sets, but can occur.
To know if the model extent (MaskMap) reaches the boundary of the meteo input set, the readmeteodata() function needs to output this information, which is then used in the downscaling2() function.
The disadvantage is that you have an additional parameter that readmeteodata() needs to output